rename cmdline.py to __main__ and move code to parser for now
