+++ src/deltabot/__main__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from .parser import main
+
+
+if __name__ == "__main__":
+    main()
